1.3.0-alpha.4 - Updated `dirt_title_screen` to `title_screen`. Added support for options_background.png when using `Dirt` Title Screen or `Legacy` UI Background, Added Shader Selection Screen Toggle Blur Button.

+++ src/main/java/com/mclegoman/perspective/client/screen/config/shaders/ShadersListWidget.java
@@ -14,7 +14,7 @@ import net.minecraft.client.gui.widget.AlwaysSelectedEntryListWidget;
 import org.jetbrains.annotations.Nullable;
 
 public class ShadersListWidget<E extends AlwaysSelectedEntryListWidget.Entry<E>> extends AlwaysSelectedEntryListWidget<ShaderListEntry> {
-	protected ShadersListWidget(int width, int height, int top, int bottom, int itemHeight, int scrollAmount) {
+	protected ShadersListWidget(int width, int height, int top, int bottom, int itemHeight, double scrollAmount) {
 		super(ClientData.CLIENT, width, height - top - bottom, top, itemHeight);
 		for (int i = 0; i <= ShaderDataLoader.getShaderAmount(); i++) {
 			this.addEntry(new ShaderListEntry(i));
